+++ CHANGELOG.md
@@ -18,6 +18,12 @@ and this project adheres to [Python Versioning](http://legacy.python.org/dev/pep
 * #### NI-ModInst
   * #### Added
   * #### Changed
+    * Device index is now on session not attribute. The correct way is now
+      ```python
+      i = 0
+      with nimodinst.Session('nidmm') as session:
+          name = session[i].device_name
+      ```
   * #### Removed
 * #### NI-Switch
   * #### Added
@@ -52,3 +58,4 @@ and this project adheres to [Python Versioning](http://legacy.python.org/dev/pep
   * #### Changed
   * #### Removed
 -->
+